pyproject.toml
[project]
name = "2025-project-diffmodels"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=1.6.0",
    "beautifulsoup4>=4.13.3",
    "bitsandbytes>=0.45.5",
    "diffusers>=0.33.1",
    "einops==0.7.0",
    "huggingface-hub>=0.30.2",
    "ip-adapter",
    "ipykernel>=6.29.5",
    "jsonlines>=4.0.0",
    "matplotlib>=3.10.1",
    "numpy>=2.2.4",
    "opencv-python>=4.11.0.86",
    "pandas==2.2.2",
    "pillow>=11.1.0",
    "qwen-vl-utils>=0.0.10",
    "requests>=2.32.3",
    "safetensors>=0.5.3",
    "scikit-learn==1.5.0",
    "timm>=0.6.12",
    "torch>=2.6.0",
    "torchvision>=0.21.0",
    "tqdm>=4.67.1",
    "transformers>=4.51.1",
    "ultralytics>=8.3.23",
    "wandb>=0.19.9",
]

[tool.uv.sources]
ip-adapter = { git = "https://github.com/tencent-ailab/IP-Adapter.git" }
